+++ src/mlpack/tests/activation_functions_test.cpp
@@ -24,6 +24,7 @@
 #include <mlpack/methods/ann/activation_functions/mish_function.hpp>
 #include <mlpack/methods/ann/activation_functions/lisht_function.hpp>
 #include <mlpack/methods/ann/activation_functions/gelu_function.hpp>
+#include <mlpack/methods/ann/activation_functions/elliot_function.hpp>
 #include <mlpack/methods/ann/activation_functions/elish_function.hpp>
 
 #include <boost/test/unit_test.hpp>
@@ -862,6 +863,26 @@ BOOST_AUTO_TEST_CASE(HardShrinkFunctionTest)
 }
 
 /**
+ * Basic test of the Elliot function.
+ */
+BOOST_AUTO_TEST_CASE(ElliotFunctionTest)
+{
+  // Calculated using PyTorch tensor.
+  const arma::colvec desiredActivations("-0.66666667 0.76190476 0.81818182 \
+                                         -0.99011858 0.5 -0.5 \
+                                          0.66666667 0.0 ");
+
+  const arma::colvec desiredDerivatives("0.36 0.32213294 0.3025 \
+                                         0.25248879 0.44444444 \
+                                         0.44444444 0.36 1.0 ");
+
+  CheckActivationCorrect<ElliotFunction>(activationData,
+                                         desiredActivations);
+  CheckDerivativeCorrect<ElliotFunction>(desiredActivations,
+                                         desiredDerivatives);
+}
+
+/**
  * Basic test of the EliSH function.
  */
 BOOST_AUTO_TEST_CASE(ElishFunctionTest)
@@ -881,7 +902,7 @@ BOOST_AUTO_TEST_CASE(ElishFunctionTest)
                                         desiredDerivatives);
 }
 
- /** 
+/** 
  * Basic test of the Soft Shrink function.
  */
 BOOST_AUTO_TEST_CASE(SoftShrinkFunctionTest)