+++ HISTORY.md
@@ -3,6 +3,8 @@
   * Added `Multi Label Soft Margin Loss` loss function for neural networks
    (#2345).
 
+  * Addition of `Elliot` activation function (#2268).
+
   * Add adaptive max pooling and adaptive mean pooling layers (#2195).
 
 ### mlpack 3.3.1